+++ src/options.js
@@ -23,6 +23,9 @@ const saveOptions = () => {
   const largeEmoteSize = document.getElementById('large-emote-size').value
   const useFrozenEmotes = document.getElementById('use-frozen-emotes').checked
   const emoteSuggestionAmount = document.getElementById('emote-suggestion-amount').value
+  const emoteFuzzyMatchingLocation = document.getElementById('emote-fuzzy-matching-location').value
+  const emoteFuzzyMatchingDistance = document.getElementById('emote-fuzzy-matching-distance').value
+  const emoteFuzzyMatchingThreshold = document.getElementById('emote-fuzzy-matching-threshold').value
 
   // eslint-disable-next-line no-undef
   chrome.storage.sync.set({
@@ -31,7 +34,10 @@ const saveOptions = () => {
     defaultEmoteSize,
     largeEmoteSize,
     useFrozenEmotes,
-    emoteSuggestionAmount
+    emoteSuggestionAmount,
+    emoteFuzzyMatchingLocation,
+    emoteFuzzyMatchingDistance,
+    emoteFuzzyMatchingThreshold
   }, () => {
     const status = document.getElementById('status')
     status.textContent = 'Options saved.'
@@ -50,7 +56,10 @@ const restoreOptions = () => {
     defaultEmoteSize: '2em',
     largeEmoteSize: '4em',
     useFrozenEmotes: false,
-    emoteSuggestionAmount: 10
+    emoteSuggestionAmount: 10,
+    emoteFuzzyMatchingLocation: 0,
+    emoteFuzzyMatchingDistance: 100,
+    emoteFuzzyMatchingThreshold: 0.6
   }, items => {
     document.getElementById('emotes-url').value = items.emotesUrl
     document.getElementById('access-key').value = items.accessKey
@@ -58,6 +67,9 @@ const restoreOptions = () => {
     document.getElementById('large-emote-size').value = items.largeEmoteSize
     document.getElementById('use-frozen-emotes').checked = items.useFrozenEmotes
     document.getElementById('emote-suggestion-amount').value = items.emoteSuggestionAmount
+    document.getElementById('emote-fuzzy-matching-location').value = items.emoteFuzzyMatchingLocation
+    document.getElementById('emote-fuzzy-matching-distance').value = items.emoteFuzzyMatchingDistance
+    document.getElementById('emote-fuzzy-matching-threshold').value = items.emoteFuzzyMatchingThreshold
   })
 }